requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==45.0.4
python-dotenv==1.0.1
